Added a Context Provider/Consumer.
The previous code creates a Provider based on a Context component that
is
passed as a prop and sets a value based on the return from the
withDataFetching HOC that is fetching all of the lists. Using the
children
prop, all of the components that will be wrapped inside the
ListsContextProvider component can retrieve the data for the value
from a
Consumer.